+++ packages/tools/utils/CHANGELOG.md
@@ -4,6 +4,14 @@
 
 - add `raiseWith` (was originally provided in `v0.10.0` but I forgot to export `src/[any-boolean]-[array-array]` in that release)
 
+### (Any -> Number) -> (Array -> Number)
+
+- add `makeAverageWith`
+
+### Array -> Number
+
+- add `arrayAverage`, `keyValueArrayAverage`
+
 
 ## `@svizzle/utils` v0.10.0